requirements.txt
certifi==2020.6.20
comtypes==1.1.10
flatbuffers==2.0
future==0.18.2
numpy~=1.21.6
onnxruntime==1.9.0
opencv-python==4.3.0.36
Pillow~=9.2.0
protobuf==3.19.1
pyasn1==0.4.8
pyperclip==1.8.2
pywin32==223
rsa==4.7.2
torch~=1.12.1+cu113
torchvision~=0.13.1+cu113
typing_extensions==4.0.0
wincertstore==0.2
WMI==1.5.1

json5~=0.9.6
matplotlib~=3.5.3